Updates kp tests to pass in MacOS env.

+++ tests/solvers/test_kp.py
@@ -21,7 +21,7 @@ from digneapy.solvers import default_kp, map_kp, miw_kp, mpw_kp
 def default_instance():
     p = np.asarray(list(range(1, 101)), dtype=np.uint32)
     w = np.asarray(list(range(1, 101)), dtype=np.uint32)
-    q = 50
+    q = np.uint64(50)
     return knapsack.Knapsack(p, w, q)
 
 
@@ -31,7 +31,7 @@ def default_large_knap():
     c = rng.integers(1e3, 1e5)
     w = rng.integers(1000, 5000, size=1000, dtype=np.uint32)
     p = rng.integers(1000, 5000, size=1000, dtype=np.uint32)
-    kp = knapsack.Knapsack(profits=p, weights=w, capacity=c)
+    kp = knapsack.Knapsack(profits=p, weights=w, capacity=c.astype(np.uint64))
 
     return kp
